Refactor: 폴더 아이콘 공통 컴포넌트로 리팩토링
